feature/24ブランチでServiceの絞り込み検索ãの実装を追加（課題24）
