+++ src/main/java/raisetech/student/management/Application.java
@@ -1,4 +1,4 @@
-package student.management.student.management;
+package raisetech.student.management;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,28 +6,15 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourses;
+import raisetech.student.management.repository.StudentRepository;
 
 @SpringBootApplication //SpringBootアプリの起動（SpringBootが使える）
-@RestController
 public class Application {
-
-  // 【講義19】StudentRepositoryというDB操作用のインターフェースを変数repositoryに注入してほしい（newしなくてよい）
-  @Autowired
-  private StudentRepository repository;
-
 
   public static void main(String[] args) {
     SpringApplication.run(Application.class, args);
-  }
-
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return repository.search();
-  }
-
-  @GetMapping("/studentCoursesList")
-  public List<StudentCourses> getStudentCoursesList() {
-    return repository.searchCourses();
   }
 
 }
