+++ src/main/java/raisetech/student/management/data/StudentCourses.java
@@ -1,6 +1,6 @@
-package student.management.student.management;
+package raisetech.student.management.data;
 
-import java.util.Date;
+import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,7 +12,7 @@ public class StudentCourses {
   private String id;
   private String studentId;
   private String courseName;
-  private Date startDate;
-  private Date endDate;
+  private LocalDateTime startDate;
+  private LocalDateTime endDate;
 
 }
