+++ src/test/java/raisetech/student/management/ApplicationTests.java
@@ -1,4 +1,4 @@
-package student.management.student.management;
+package raisetech.student.management;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
